split node iterators into separate file
