+++ components/script/dom/node/focus.rs
@@ -10,10 +10,11 @@ use script_bindings::inheritance::Castable;
 use script_bindings::root::DomRoot;
 
 use crate::dom::document::focus::{FocusableArea, FocusableAreaKind};
+use crate::dom::node::iterators::TreeIterator;
 use crate::dom::types::{
     Element, HTMLDialogElement, HTMLIFrameElement, HTMLSlotElement, ShadowRoot,
 };
-use crate::dom::{Document, Node, NodeTraits, ShadowIncluding, TreeIterator};
+use crate::dom::{Document, Node, NodeTraits, ShadowIncluding};
 
 /// <https://html.spec.whatwg.org/multipage/#focus-navigation-scope-owner>
 ///
