+++ components/script/xpath.rs
@@ -28,7 +28,8 @@ use crate::dom::comment::Comment;
 use crate::dom::document::Document;
 use crate::dom::element::Element;
 use crate::dom::element::attributes::storage::AttributeStorage;
-use crate::dom::node::{Node, NodeTraits, PrecedingNodeIterator, ShadowIncluding};
+use crate::dom::iterators::PrecedingNodeIterator;
+use crate::dom::node::{Node, NodeTraits, ShadowIncluding};
 use crate::dom::processinginstruction::ProcessingInstruction;
 use crate::dom::text::Text;
